moved & merged files to the custom_module
