+++ nodes/firmware/custom-modules/battery_voltage_monitor/Makefile
@@ -1,0 +1,4 @@
+MODULE := battery_voltage_monitor
+SRC := $(MODULE).c
+
+include $(RIOTBASE)/Makefile.base
