+++ nodes/firmware/custom-modules/battery_voltage_monitor/Makefile.dep
@@ -1,0 +1,2 @@
+USEMODULE += periph_adc
+USEMODULE += periph_gpio
