+++ nodes/firmware/custom-modules/battery_voltage_monitor/Makefile.include
@@ -1,0 +1,3 @@
+# Use an immediate variable to evaluate `MAKEFILE_LIST` now
+USEMODULE_INCLUDES_battery_voltage_monitor := $(LAST_MAKEFILEDIR)/include
+USEMODULE_INCLUDES += $(USEMODULE_INCLUDES_battery_voltage_monitor)
